encode_decode: rename files for generated param parsing
