+++ .gitignore
@@ -86,6 +86,13 @@ providers/common/include/prov/der_wrap.h
 providers/common/include/prov/der_sm2.h
 providers/common/include/prov/der_ml_dsa.h
 providers/common/include/prov/der_hkdf.h
+providers/implementations/encode_decode/decode_der2key.c
+providers/implementations/encode_decode/decode_epki2pki.c
+providers/implementations/encode_decode/decode_pem2der.c
+providers/implementations/encode_decode/decode_pvk2key.c
+providers/implementations/encode_decode/decode_spki2typespki.c
+providers/implementations/encode_decode/encode_key2any.c
+providers/implementations/encode_decode/encode_key2ms.c
 providers/implementations/kdfs/argon2.c
 providers/implementations/kdfs/hkdf.c
 providers/implementations/kdfs/hmacdrbg_kdf.c
