+++ build.info
@@ -48,6 +48,13 @@ DEPEND[]=include/openssl/asn1.h \
          include/openssl/x509_acert.h \
          include/openssl/x509_vfy.h \
          include/crypto/bn_conf.h include/crypto/dso_conf.h \
+         providers/implementations/encode_decode/decode_der2key.c \
+         providers/implementations/encode_decode/decode_epki2pki.c \
+         providers/implementations/encode_decode/decode_pem2der.c \
+         providers/implementations/encode_decode/decode_pvk2key.c \
+         providers/implementations/encode_decode/decode_spki2typespki.c \
+         providers/implementations/encode_decode/encode_key2any.c \
+         providers/implementations/encode_decode/encode_key2ms.c \
          providers/implementations/kdfs/argon2.c \
          providers/implementations/kdfs/hkdf.c \
          providers/implementations/kdfs/hmacdrbg_kdf.c \
@@ -109,7 +116,14 @@ GENERATE[include/openssl/x509_vfy.h]=include/openssl/x509_vfy.h.in
 GENERATE[include/crypto/bn_conf.h]=include/crypto/bn_conf.h.in
 GENERATE[include/crypto/dso_conf.h]=include/crypto/dso_conf.h.in
 
-DEPEND[providers/implementations/kdfs/argon2.c \
+DEPEND[providers/implementations/encode_decode/decode_der2key.c \
+       providers/implementations/encode_decode/decode_epki2pki.c \
+       providers/implementations/encode_decode/decode_pem2der.c \
+       providers/implementations/encode_decode/decode_pvk2key.c \
+       providers/implementations/encode_decode/decode_spki2typespki.c \
+       providers/implementations/encode_decode/encode_key2any.c \
+       providers/implementations/encode_decode/encode_key2ms.c \
+       providers/implementations/kdfs/argon2.c \
        providers/implementations/kdfs/hkdf.c \
        providers/implementations/kdfs/hmacdrbg_kdf.c \
        providers/implementations/kdfs/kbkdf.c \
@@ -137,6 +151,20 @@ DEPEND[providers/implementations/kdfs/argon2.c \
        providers/implementations/digests/digestcommon.c \
        providers/implementations/digests/sha3_prov.c \
        include/openssl/core_names.h]=util/perl|OpenSSL/paramnames.pm
+GENERATE[providers/implementations/encode_decode/decode_der2key.c]=\
+    providers/implementations/encode_decode/decode_der2key.c.in
+GENERATE[providers/implementations/encode_decode/decode_epki2pki.c]=\
+    providers/implementations/encode_decode/decode_epki2pki.c.in
+GENERATE[providers/implementations/encode_decode/decode_pem2der.c]=\
+    providers/implementations/encode_decode/decode_pem2der.c.in
+GENERATE[providers/implementations/encode_decode/decode_pvk2key.c]=\
+    providers/implementations/encode_decode/decode_pvk2key.c.in
+GENERATE[providers/implementations/encode_decode/decode_spki2typespki.c]=\
+    providers/implementations/encode_decode/decode_spki2typespki.c.in
+GENERATE[providers/implementations/encode_decode/encode_key2any.c]=\
+    providers/implementations/encode_decode/encode_key2any.c.in
+GENERATE[providers/implementations/encode_decode/encode_key2ms.c]=\
+    providers/implementations/encode_decode/encode_key2ms.c.in
 GENERATE[providers/implementations/kdfs/argon2.c]=\
     providers/implementations/kdfs/argon2.c.in
 GENERATE[providers/implementations/kdfs/hkdf.c]=\
