+++ providers/implementations/encode_decode/decode_msblob2key.c
@@ -28,7 +28,7 @@
 #include "crypto/rsa.h"
 #include "prov/bio.h"
 #include "prov/implementations.h"
-#include "endecoder_local.h"
+#include "prov/endecoder_local.h"
 
 struct msblob2key_ctx_st;            /* Forward declaration */
 typedef void *b2i_of_void_fn(const unsigned char **in, unsigned int bitlen,
