+++ providers/implementations/encode_decode/encode_key2any.c.in
@@ -38,7 +38,7 @@
 #include "prov/bio.h"
 #include "prov/provider_ctx.h"
 #include "prov/der_rsa.h"
-#include "endecoder_local.h"
+#include "prov/endecoder_local.h"
 #include "ml_dsa_codecs.h"
 #include "ml_kem_codecs.h"
