+++ providers/implementations/encode_decode/encode_key2text.c
@@ -30,9 +30,9 @@
 #include "prov/bio.h"
 #include "prov/implementations.h"
 #include "internal/encoder.h"
-#include "endecoder_local.h"
-#include "ml_dsa_codecs.h"
-#include "ml_kem_codecs.h"
+#include "prov/endecoder_local.h"
+#include "prov/ml_dsa_codecs.h"
+#include "prov/ml_kem_codecs.h"
 
 DEFINE_SPECIAL_STACK_OF_CONST(BIGNUM_const, BIGNUM)
