+++ providers/implementations/include/prov/ml_kem_codecs.h
@@ -15,7 +15,7 @@
 #  include <openssl/e_os2.h>
 #  include "crypto/ml_kem.h"
 #  include "prov/provider_ctx.h"
-#  include "ml_common_codecs.h"
+#  include "prov/ml_common_codecs.h"
 
 __owur
 ML_KEM_KEY *ossl_ml_kem_d2i_PUBKEY(const uint8_t *pubenc, int publen,
